Move metrics related functions into upgrade_metrics.go
